+++ src/app.module.ts
@@ -3,7 +3,6 @@ import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { ConfigModule, ConfigService } from '@nestjs/config';
 import { APP_INTERCEPTOR } from '@nestjs/core';
-import { UserModule } from './user/user.module';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { MerchantModule } from './merchant/merchant.module';
 import { IdentityModule } from './identity/identity.module';
@@ -31,7 +30,6 @@ import { MemberModule } from './member/member.module';
       inject: [ConfigService],
     }),
 
-    UserModule,
     MerchantModule,
     IdentityModule,
     MemberModule,
