+++ src/identity/identity.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { IdentityRepository } from './identity.repository';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { Identity } from './identity.entity';
+import { Identity } from './entities/identity.entity';
 import { IdentityService } from './identity.service';
 import { IdentityController } from './identity.controller';
