+++ src/member/member.repositiry.ts
@@ -1,4 +1,4 @@
-import { Identity } from 'src/identity/identity.entity';
+import { Identity } from 'src/identity/entities/identity.entity';
 import { Repository } from 'typeorm';
 
 export class IdentityRepository extends Repository<Identity> {}
