+++ src/merchant/merchant.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { Merchant } from './merchant.entity';
+import { Merchant } from './entities/merchant.entity';
 import { MerchantRepository } from './merchant.repository';
 import { MerchantController } from './merchant.controller';
 import { MerchantService } from './merchant.service';
