+++ src/merchant/merchant.repository.ts
@@ -1,4 +1,4 @@
 import { Repository } from 'typeorm';
-import { Merchant } from './merchant.entity';
+import { Merchant } from './entities/merchant.entity';
 
 export class MerchantRepository extends Repository<Merchant> {}
