--- src/user/user.module.ts
@@ -1,7 +1,0 @@
-import { Module } from '@nestjs/common';
-import { UserController } from './user.controller';
-
-@Module({
-  controllers: [UserController]
-})
-export class UserModule {}
